Add v1_sync_helper client
This client is not actually authenticated-to by v1-sync-helper, but that
service, which impersonates authenticated users via its own Heimdall-JWT
generation, may use this client ID as the authenticated principal in a
bearer token if there is no relevant principal to impersonate for a
given data update.

Also update Chart dependencies and bump chart version.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>